search_ui/requirements.txt
streamlit
fastapi
uvicorn
requests
supabase
sentence-transformers
python-dotenv
torch # Or tensorflow/flax depending on your backend preference
beautifulsoup4
